backend/scribe_eye_pro/requirements.txt
fastapi[all]
motor
pydantic[email]
python-dotenv
pinecone-client
langchain
langchain-google-genai
langchain-community
langchain-pinecone